requirements.txt
absl-py==2.1.0
altair==5.2.0
anaconda-client==1.2.2
annotated-types==0.6.0
ansi2html==1.8.0
appdirs==1.4.4
astunparse==1.6.3
atomicwrites==1.4.0

autograd==1.6.2
autograd-gamma==0.5.0
autokeras==1.0.20

autoviz==0.1.732
backports.weakref==1.0.post1

blinker==1.7.0

brotlipy==0.7.0
cachetools==5.3.2

clusteval==2.2.1

colourmap==1.1.16

Cython==0.29.37

daal4py==2023.1.1
dash==2.14.1
dash-bootstrap-components==1.3.1
dash-colorscales==0.0.4
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0

datazets==0.1.9
deap==1.4.1

dm-tree==0.1.8

dtale==3.7.0
emoji==2.8.0

et-xmlfile==1.1.0

feature-engine==1.6.2

FLAML==2.1.2
Flask-Compress==1.14
flask-ngrok==0.0.25
flatbuffers==24.3.7
fonttools==4.25.0
formulaic==1.0.1
gast==0.5.4

gitdb==4.0.11
GitPython==3.1.41

google-pasta==0.2.0

grpcio==1.62.1
h2o==3.46.0.1
h5py==3.10.0

htmlmin==0.1.12
hvplot==0.7.3

ImageHash==4.3.1

imblearn==0.0

importlib-resources==6.1.0

inflection==0.5.1

interface-meta==1.3.0

jsonpointer==2.1

jupyter==1.0.0

kaleido==0.2.1
keras==3.0.5
keras-tuner==1.4.7

kneed==0.8.5
kt-legacy==1.0.5


libclang==18.1.1

lifelines==0.28.0
lightgbm==4.3.0

loguru==0.7.2

missingno==0.5.2

ml-dtypes==0.3.2

multimethod==1.10

munkres==1.1.4

namex==0.0.7

openpyxl==3.0.10
opt-einsum==3.3.0

pandas-dq==1.28

patsy==0.5.3
pep8==1.7.1


ply==3.11

protobuf==4.25.2


py-AutoClean==1.1.3

pyamg==5.0.1

pyasn1-modules==0.2.8

pydeck==0.8.1b0
PyDispatcher==2.0.5

pyls-spyder==0.4.0
PyMySQL==1.1.0

pypickle==1.1.0
PyQt5==5.15.10

PyQtWebEngine==5.15.6


retrying==1.3.4

scatterd==1.3.7

scikit-learn-intelex==2023.1.1
scipy==1.11.4

sidetable==0.9.1

sklearn-pandas==2.2.0
smmap==5.0.1

squarify==0.4.3
SQLAlchemy== 2.0.25
stopit==1.1.2
streamlit==1.30.0
strsimpy==0.2.1
sweetviz==2.2.1

tangled-up-in-unicode==0.2.0
TBB==2021.11.0

tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==0.0.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0

textblob==0.17.1

TPOT==0.12.2

tzlocal==5.2

update-checker==0.18.0

validators==0.22.0
visions==0.7.4

webencodings==0.5.1

win32-setctime==1.1.0
wordcloud==1.9.2

xgboost==2.0.1
xlrd==2.0.1

zstandard==0.19.0
